fix(evaluation): Run normalization and analysis in separate containers
Initially wanted to run the analysis script inside the original fsl container
to not complicate this unnecessarily. However, the FSL docker container runs
Ubuntu 14 and python 3.4 and it was not possible to upgrade this to a version
which support nibabel and nilearn. Therefore, we use an extra container with
a newer python image just for the analysis.