feat(SIRF): Save results via SIRF builtin Nifti output class
Didn't know about the SIRF builtin solution yet but the manual
wrangling with the data caused incorrect MNI normalization downstream.

The problem was likely due to some incorrect settings in the resulting headers,
however, due to the builtin class we do not have to care about it and
it works now.

Because the class only supports single frames and no 4D images,
we need to additionally merge all result files into one.